Refactor internal logic of ViewModel creation.
If creating a ViewModel with the default arguments, use HasDefaultViewModelProviderFactory interface.
